Add global reusable confirmation dialog and small fix of recipes-view component

+++ src/simply-recipes/src/app/recipes/recipes-view/recipes-view.component.ts
@@ -1,7 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
 import { RecipesService } from 'src/app/services/recipes.service';
-import { ErrorDialogComponent } from 'src/app/shared/error-dialog/error-dialog.component';
+import { ConfirmDialogModel, ConfirmationDialogComponent } from 'src/app/shared/dialogs/confirmation-dialog/confirmation-dialog.component';
+import { ErrorDialogComponent } from 'src/app/shared/dialogs/error-dialog/error-dialog.component';
 import { Difficulty } from 'src/app/shared/enums/difficulty';
 import { IRecipeDetails } from 'src/app/shared/interfaces/recipes/recipe-details';
 
@@ -24,14 +25,22 @@ export class RecipesViewComponent implements OnInit {
   }
 
   onDeleteHandler(recipeId: number): void {
-    this.recipesService.removeRecipe(recipeId).subscribe({
-      next: () => {
-        this.getUserRecipes();
-       }, // TODO: Open success dialog
-      error: (err: string) => {
-        this.dialog.open(ErrorDialogComponent, {
-          data: {
-            message: err
+    const message = 'Are you sure you want to delete this recipe?';
+    const dialogData = new ConfirmDialogModel('Confirmation', message);
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      data: dialogData
+    });
+
+    dialogRef.afterClosed().subscribe(dialogResult => {
+      if (dialogResult) {
+        this.recipesService.removeRecipe(recipeId).subscribe({
+          next: () => {
+            this.getUserRecipes();
+          },
+          error: (err: string) => {
+            this.dialog.open(ErrorDialogComponent, {
+              data: { message: err }
+            });
           }
         });
       }
